requirements.txt
streamlit>=1.32.0
google-generativeai>=0.5.0
sentence-transformers>=2.7.0
torch>=2.2.0
